# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "{{.Description}}",
        "title": "{{.Title}}",
        "contact": {},
        "license": {},
        "version": "{{.Version}}"
    },
    "host": "{{.Host}}",
    "basePath": "{{.BasePath}}",
    "paths": {
        "/v1/classic/detail/{type}/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classic"
                ],
                "summary": "获取期刊详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "[{\"id\": 1,\"image\": \"images/music.7.png\",\"content\": \"无人问我粥可温 风雨不见江湖人\", \"title\": \"月之门《枫华谷的枫林》\",\"fav_nums\": 145,\"type\": 200,\"url\": \"http://music.163.com/song/media/outer/url?id=393926.mp3\"}]",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/classic/favor/{type}/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classic"
                ],
                "summary": "用户是否喜欢一个期刊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Type",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "[{\"id\": 1,\"image\": \"images/music.7.png\",\"content\": \"无人问我粥可温 风雨不见江湖人\", \"title\": \"月之门《枫华谷的枫林》\",\"fav_nums\": 145,\"type\": 200,\"url\": \"http://music.163.com/song/media/outer/url?id=393926.mp3\"}]",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/classic/favors": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classic"
                ],
                "summary": "获取用户喜欢的期刊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"content\": \"人生不能像做菜，把所有的料准备好才下锅\",\"favnums\": 0,\"id\": 7,\"image\": \"images/movie.8.png\",\"index\": 8,\"like_status\": true,\"pubdate\": \"2019-04-05T17:12:04+08:00\",\"title\": \"李安《饮食男女 》\",\"type\": 100}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/classic/latest": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classic"
                ],
                "summary": "获取最新期刊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"content\": \"人生不能像做菜，把所有的料准备好才下锅\",\"favnums\": 0,\"id\": 7,\"image\": \"images/movie.8.png\",\"index\": 8,\"like_status\": true,\"pubdate\": \"2019-04-05T17:12:04+08:00\",\"title\": \"李安《饮食男女 》\",\"type\": 100}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/classic/next/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classic"
                ],
                "summary": "获取当前期刊的下一个期刊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/classic/previous/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Classic"
                ],
                "summary": "获取当前期刊的上一个期刊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/like/": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Like"
                ],
                "summary": "期刊点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "likeParam",
                        "name": "likeParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ParamFavorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/like/cancel": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Like"
                ],
                "summary": "期刊取消点赞",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "likeParam",
                        "name": "likeParam",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.ParamFavorReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        },
        "/v1/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "邮箱方式注册用户",
                "parameters": [
                    {
                        "description": "邮箱方式注册用户",
                        "name": "service",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.CreateResponse"
                        }
                    }
                }
            }
        },
        "/v1/user/token": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "获取token",
                "parameters": [
                    {
                        "description": "token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.TokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"token\": \"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1Njc3Njk4MDYsImlhdCI6MTU2Nzc2MjYwNiwiaWQiOjEsIm5iZiI6MTU2Nzc2MjYwNiwic2NvcGUiOjh9.-VlL6oAa8mMD2Wd0Os1in1V5T9sdcwv6OCupihZKZNY\"}",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/model.Token"
                        }
                    }
                }
            }
        },
        "/v1/user/token/verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "验证Token合法性",
                "parameters": [
                    {
                        "type": "string",
                        "description": "x-auth-token",
                        "name": "x-auth-token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/service.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.CreateResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
        "model.ParamFavorReq": {
            "type": "object",
            "required": [
                "id",
                "type"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "model.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "model.TokenRequest": {
            "type": "object",
            "required": [
                "account"
            ],
            "properties": {
                "account": {
                    "type": "string"
                },
                "secret": {
                    "type": "string"
                },
                "type": {
                    "type": "integer"
                }
            }
        },
        "service.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}